refactor(teamSelection): extract useTeamSelection controller hook
Move the page's entire state machine — 9 useState, the region/country/
competition sync effects, all derived selectors, and the region/
competition toggle + confirm handlers — into a useTeamSelection hook.
TeamSelection.tsx becomes a 172-line orchestrator (from 564) that wires
the hook outputs to the scope panel, grid, and sidebar.

The hook is called unconditionally before the null-gameState guard,
which regularizes a pre-existing Rules-of-Hooks violation (teamGroups
useMemo + team-sync effect previously ran after an early return). The
only non-verbatim change is null-safety on the post-guard derivations
(gameState?.teams ?? [] etc.), behavior-identical on the happy path.

Adds useTeamSelection.test.ts with renderHook coverage of the toggle
constraint logic (home-region lock, region enable, mandatory-region
block, dependent-competition cascade, auto-enable regions, locked-comp
guard, plain disable) — the orchestration the helper tests did not reach.
